+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -125,6 +125,12 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`bugprone-casting-through-void
+  <clang-tidy/checks/bugprone/casting-through-void>` check.
+
+  Detects usage of ``static_cast`` pointer to the other pointer throght
+  ``static_cast`` to ``void *`` in C++ code.
+
 - New :doc:`bugprone-compare-pointer-to-member-virtual-function
   <clang-tidy/checks/bugprone/compare-pointer-to-member-virtual-function>` check.